changed encoding of all localization files. added one special char in comment field to allow editors recognize utf-8 encodingof files that do only use ASCII characters
-------------------------------------------------------------------------------
i got a little problem with encoding of localization files.

i will try to explain:
files do not have a property like "encoding=". editors take a look at file and check how containing characters are encoded. if there is not a single special character like the german umlauts every editor expects it to be ascii -> editor opens text as ascii. if then the file is edited and german umlauts are added they are saved as ASCII -> problem.

i got some kind of fix for that.

every localization file has to contain at least 1 special character