TICKET 4929: DB Access Error while importing requirements from xml file
If the user selected a project from the tree structure, the req spec id
argument would be 0. And if his/her XML had no req spec id, the
requirements would be created in the database without a requirement
specification parent.

This could lead to inconsistencies, or DB errors. This patch fixes the
bug by introducing a check for the situation described above. As there
is a new error message, a new string has been created in the US locale
files.