TICKET-4576: Unable to import requirement using CSV file
The CSV parser reported a syntax error that wasn't being used. Now,
once this message happens, the user will see a message saying that
the import was done, but with errors.

There are cases when a CSV file may contain only one line with syntax
error. The current implementation parses the file until it reaches the
EOF, independently if errors occurred or not.