TICKET-4977: Advisory ID: HTB23088 - CSRF issue
This commit incorporates two scripts from Mantis into TestLink. Both
files had to be massaged to TestLink API and needs, and they are
responsible for generating and validating security tokens, put in
hidden fields in TestLink forms.

If the token is not present, is expired, or does not match the
token stored in the session, then a message is displayed in the
user interface.

Further changes are required for the remaining forms in TestLink.