Fixed critical issue when editing test cases or creating req specs.
The ajaxCheck case block verifies information from the session and sends
it back to the caller. This way, it is possible to validate if the session
is still active or not. This is used for avoiding loss of data when the
session has expired (a popup is displayed in this case).

We have removed the CSRF token validation from this case block. Although it
allows users to submit without verifying the token, this is a simple GET call
that only retrieves information if the session is active or not, and does not
change any data, so it is less harmful then other functions (like changing
password).

A basic test was executed, logging normally, then creating a req spec and
finally editing a test case.