Added Plugin Enable/Disable UI.
The Plugin enable/disable UI has 2 tables. The top table shows the list of enabled plugins that
can be uninstalled and the bottom table contains the list of new plugins that are yet to be
installed. The details that are shown in the page are the Plugin Name, Description and the Version.

Major changes:
* Added new permission (mgt_plugins) that is only added to the "admin" role for now. This "rights"
  is required for a user to see the plugin management page. This will also appear in the "System
  rights" section of the Roles page.
* Migration scripts for the following:
  + Added the mgt_plugins right
  + Added the right to the "admin" role
  + Default tables required for plugins to work
  This is done for all supported DBs.
* Additional functions added in tlPlugin.class.php so actual plugins have the opportunity to
  add custom code when a plugin is activated ("install" function) or deactived("uninstall" function)
* Added Language bindings only for en_GB