BUG FIX for jirarestInterface.class.php
In function getCustomFieldsAttribute($name,$objCFSet)

Line   761:   $cfJIRAID = $cf['customfield']; 
It should be 'customfieldId', otherwise the customfields are not gonna work using the Cfgtemplate .
I was getting the error "addIssue:Failure:JIRA Message: _empty_ => Field '' cannot be set. It is not on the appropriate screen, or unknown." because of this not retrieving correctly the customfields.
This little change, resolved the bug.