New config session_path
If more than one testlink is used on a webserver, it is possible now to
specify path to save sessions. In Default the new session_path will not
be used. If session_path is used the session files from php will be
saved to that path. So multiple testlink instances can be used at the
same time.