XMLRPC: Allow to fetch custom fields for testplan testcases
We use custom fields to match testlink entries to automated testcases.
At the moment the reporting tool has to place one request per testcase
in a testplan to get the custom field value, which is very slow.

With this we allow the user to request custom fields when he asks for
all testcases in a testplan. the parameter is "customfields" and it can
be either "true" for all custom fields or an array with the requested
custom field names.