TICKET 0009079: Add the ability to use the execution note template when linking with an existing issue
We can define a template for an execution note ($tlCfg->execution_template->notes) to automatically populate some contextual information.
This template is used when creating a new external issue from a Testlink bug (using the 'Create issue' icon).
But it is not used when we want to link a Testlink bug to an existing external issue (using the 'Link existent issue' icon).

We now also use this template in the case of linking.