0009126: InlineImage in steps are not retained when copying a Test Case (#324)
When we copy a Test Case whose steps used InlineImage references, we retrieve images in the copy but not referenced by InlineImage tags. They are rendered with Testlink attachment download links.

Whereas, InlineImages in Summary and Precondition are correctly retained.

InlineImages will be retained in steps data as in Test Case fields.